back/requirements.txt
django
djangorestframework
django-boost
Pillow
mysqlclient
pymysql